`ShutdownMode` has been removed. The only shutdown mode supported now is an equivalent of `ShutdownMode.Reload`
also in this change:
- fixed Python derived types not being decrefed when an instance is deallocated
- reduced time and amount of storage needed for runtime reload
- removed circular reference loop between Type <-> ConstructorBinding(s)
+ exposed Runtime.TryCollectingGarbage

+++ src/runtime/interop.cs
@@ -61,7 +61,7 @@ namespace Python.Runtime
     /// </summary>
     // Py_TPFLAGS_*
     [Flags]
-    public enum TypeFlags: int
+    public enum TypeFlags: long
     {
         HeapType = (1 << 9),
         BaseType = (1 << 10),
